Update: Enhance through refactoring hooks files
