+++ src/hooks/use-user.ts
@@ -1,4 +1,4 @@
-// REVIEWED - 08
+// REVIEWED - 09
 
 import { useMutation, useQuery } from "@tanstack/react-query";
 import { useRouter, useSearchParams } from "next/navigation";
@@ -6,10 +6,9 @@ import { toast } from "sonner";
 import { z } from "zod";
 
 import { getAuth, signIn, signOut, signUp } from "@/actions/auth";
-import { queryClient } from "@/app/(app)/providers";
-import { messages } from "@/lib/errors";
+import { messages } from "@/lib/messages";
+import { httpSafeExecute } from "@/lib/network";
 import { SignInSchema, SignUpSchema } from "@/lib/schemas/auth";
-import { httpSafeExecute } from "@/lib/utils";
 
 export const useUser = function useUser() {
   const router = useRouter();
@@ -45,7 +44,7 @@ export const useUser = function useUser() {
       }
 
       toast.success(messages.actions.auth.signIn.success);
-      queryClient.setQueryData(["user"], response.data.user);
+      refetch();
       router.push(redirectParam);
     },
   });
@@ -62,14 +61,14 @@ export const useUser = function useUser() {
       }
 
       toast.success(messages.actions.auth.signUp.success);
-      queryClient.setQueryData(["user"], response.data.user);
+      refetch();
       router.push(redirectParam);
     },
   });
 
   const signOutMutation = useMutation({
     mutationFn: async () => {
-      const response = await httpSafeExecute<{ message: string }, string>(
+      const response = await httpSafeExecute<{ message: string }>(
         fetch("/api/users/logout", {
           method: "POST",
           credentials: "include",
@@ -95,7 +94,7 @@ export const useUser = function useUser() {
       await signOut();
 
       toast.success(messages.actions.auth.signOut.success);
-      queryClient.setQueryData(["user"], null);
+      refetch();
       router.refresh();
     },
   });
